+++ vmalert.sh
@@ -36,8 +36,12 @@ fi
 ########################################################################################################################
 
 SERVICE_NAME="vmalert.service"
-TMP_DIR=/tmp/vmutils
+
+if [ -z "$TMP_DIR" ]; then
+  TMP_DIR=/tmp/vmutils
+fi
 APP_SOURCE_DIR="${TMP_DIR}"
+
 
 if [[ "$VERSION" != "$CURRENT_VERSION" ]]
 then
@@ -47,6 +51,7 @@ then
 ########################################################################################################################
   VERSION="${VERSION}" \
   TMP_DIR="${TMP_DIR}" \
+  KEEP_SOURCE="$KEEP_SOURCE" \
     tools/vmutils.sh
 ########################################################################################################################
 
@@ -61,7 +66,9 @@ fi
 
 SERVICE_NAME="$SERVICE_NAME" SERVICE_STATUS="$SERVICE_STATUS" "tools/systemd.sh"
 
-[[ -d "${TMP_DIR}" ]] && sudo rm -rf "${TMP_DIR}"
+if [ -z "$KEEP_SOURCE" ]; then
+  [[ -d "${TMP_DIR}" ]] && sudo rm -rf "${TMP_DIR}"
+fi
 
 echo ""
 echo -e "${GREEN}Installation completed${NC}"
